+++ src/context/contracts-app/contractContextFactory.tsx
@@ -3,7 +3,7 @@ import { useIsMounted } from 'usehooks-ts';
 
 import { useEthersContext } from '~~/context';
 import { checkEthersOverride } from '~~/functions';
-import { defaultHookOptions, TEthersAdaptor, TTypechainContractConnectorList, AppContractDefinitions } from '~~/models';
+import { defaultHookOptions, TEthersAdaptor, TTypechainContractConnectorList } from '~~/models';
 
 // import { AppContractDefinitions } from '~~/models/AppContractDefinitions';
 
@@ -11,30 +11,34 @@ export interface IContractsContextProps {
   ethersContextKey?: string | undefined;
 }
 
+export interface IAppContractsDefinitions<GContractNames extends string> {
+  getmonkeys: GContractNames;
+}
+
 export type TContractDispatch<GContractNames extends string> = {
   setAppContractConnectors: (appContractConnectors: TTypechainContractConnectorList<GContractNames>) => void;
 };
 
-export type TContractState<GContractNames extends string, GTypedContracts> = {
-  appcontractDefinitions: AppContractDefinitions<GContractNames, GTypedContracts>;
+export type TContractState<GContractNames extends string> = {
+  appContractDefinitions: IAppContractsDefinitions<GContractNames>;
 };
 
-export const contractContextFactory = <GContractNames extends string, GTypedContracts>(): {
+export const contractAppContextFactory = <GContractNames extends string, GTypedContracts>(): {
   ContractsContext: FC<PropsWithChildren<IContractsContextProps>>;
   ContractsDispatchContext: React.Context<TContractDispatch<GContractNames> | undefined>;
-  ContractsStateContext: React.Context<TContractState<GContractNames, GTypedContracts> | undefined>;
+  ContractsStateContext: React.Context<TContractState<GContractNames> | undefined>;
   useContractsDispatchContext: () => TContractDispatch<GContractNames> | undefined;
-  useContractsStateContext: () => TContractState<GContractNames, GTypedContracts> | undefined;
+  useContractsStateContext: () => TContractState<GContractNames> | undefined;
 } => {
   const ContractsDispatchContext = createContext<TContractDispatch<GContractNames> | undefined>(
     undefined as TContractDispatch<GContractNames> | undefined
   );
-  const ContractsStateContext = createContext<TContractState<GContractNames, GTypedContracts> | undefined>(undefined);
+  const ContractsStateContext = createContext<TContractState<GContractNames> | undefined>(undefined);
 
   const useContractsDispatchContext = (): TContractDispatch<GContractNames> | undefined => {
     return useContext(ContractsDispatchContext) as TContractDispatch<GContractNames>;
   };
-  const useContractsStateContext = (): TContractState<GContractNames, GTypedContracts> | undefined => {
+  const useContractsStateContext = (): TContractState<GContractNames> | undefined => {
     return useContext(ContractsStateContext);
   };
 
@@ -55,17 +59,18 @@ export const contractContextFactory = <GContractNames extends string, GTypedCont
     });
 
     const isMounted = useIsMounted();
-    const [state, setState] = useState<TContractState<GContractNames, GTypedContracts>>();
+    const [state, setState] = useState<TContractState<GContractNames>>();
     const [dispatchValue, setDispatchValue] = useState<TContractDispatch<GContractNames>>();
 
     const setAppContractConnectors = useCallback(
       async (appContractConnectors: TTypechainContractConnectorList<GContractNames>) => {
-        const contractDefinitions = await AppContractDefinitions.connectToAllContractsReducer<
-          GContractNames,
-          GTypedContracts
-        >(ethersAdaptor, appContractConnectors);
+        const contractDefinitions: IAppContractsDefinitions<GContractNames> = { getmonkeys: 'DAI' as GContractNames };
+        // await AppContractsDefinitions.connectToAllContractsReducer<GContractNames, GTypedContracts>(
+        //   ethersAdaptor,
+        //   appContractConnectors
+        // );
         if (isMounted()) {
-          setState({ appcontractDefinitions: contractDefinitions });
+          setState({ appContractDefinitions: contractDefinitions });
         }
       },
       [ethersAdaptor, isMounted]
