+++ src/context/ethers-app/connectors/StaticJsonRpcProviderConnector.ts
