+++ src/context/index.ts
@@ -1,4 +1,4 @@
-export * from './BlockNumberContext';
-export * from './EthersAppContext';
-export * from './connectors';
-export * from './contractContextFactory';
+export * from './ethers-app/BlockNumberContext';
+export * from './ethers-app/EthersAppContext';
+export * from './ethers-app/connectors';
+export * from './contracts-app/contractContextFactory';
