+++ src/models/index.ts
@@ -4,4 +4,4 @@ export * from './providerTypes';
 export * from './hookOptions';
 export * from './ethersContextTypes';
 export * from './typechainContractTypes';
-export * from './AppContractDefinitions';
+export * from '../context/contracts-app/AppContractDefinitions';
